pyproject.toml
[tool.poetry]
name = "benchmark-aleph-alpha-document-question-answering"
version = "0.1.0"
description = ""
authors = ["Marc Mezger <marc.mezger@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.10"
pytesseract = "^0.3.10"
aleph-alpha-client = "^2.13.0"
pdf2image = "^1.16.2"
python-dotenv = "^0.21.1"
pandas = "^1.5.3"


[tool.poetry.group.dev.dependencies]
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[settings]
known_third_party = ["PIL", "aleph_alpha_client", "dotenv", "pandas", "pdf2image", "pytesseract"]
